high level: Split single file into multiple
high_level.py becomes high_level/{ml,dataflow,source}.py

Fixes: #1170
Related: #662
