+++ model/pytorch/tests/test_pytorchnet.py
@@ -7,7 +7,7 @@ import asyncio
 from dffml.cli.cli import CLI
 from dffml.util.net import cached_download_unpack_archive
 from dffml.util.asynctestcase import AsyncTestCase
-from dffml.high_level import train, accuracy, predict
+from dffml.high_level.ml import train, accuracy, predict
 from dffml import Features, Feature, DirectorySource
 from dffml_model_pytorch import PyTorchNeuralNetwork
 from dffml_model_pytorch.utils import CrossEntropyLossFunction
