+++ model/scikit/tests/test_scikit.py
@@ -3,7 +3,7 @@ import tempfile
 import numpy as np
 
 from dffml.record import Record
-from dffml.high_level import accuracy
+from dffml.high_level.ml import accuracy
 from dffml.source.source import Sources
 from dffml.source.memory import MemorySource, MemorySourceConfig
 from dffml.feature import Feature, Features
