+++ scripts/docs/templates/api/high_level.rst
@@ -3,14 +3,14 @@ High Level
 
 Example usage can be found in the :doc:`/quickstart/model`.
 
-.. autofunction:: dffml.high_level.train
+.. autofunction:: dffml.high_level.ml.train
 
-.. autofunction:: dffml.high_level.accuracy
+.. autofunction:: dffml.high_level.ml.accuracy
 
-.. autofunction:: dffml.high_level.predict
+.. autofunction:: dffml.high_level.ml.predict
 
-.. autofunction:: dffml.high_level.save
+.. autofunction:: dffml.high_level.source.save
 
-.. autofunction:: dffml.high_level.load
+.. autofunction:: dffml.high_level.source.load
 
-.. autofunction:: dffml.high_level.run
+.. autofunction:: dffml.high_level.dataflow.run
